Created test file with reasons/README
Another thing to add to the list is to create a “test” VMR.xslx. It should only contain ~ 10 rows. Choose rows to meet these criteria:
just a simple accession
a list of accessions with NO segment names
a list of accessions WITH segment names
include both an E and >1 A record for the same species
include one Poxviridae species (large genome size)
include “Kayvirus G1” and another species from Keyvirus genus (they’ve given us blast problems in the past)

And add that .xlsx file to git.

It is good practice (and will save you a lot of time) to have a small dataset, with representative items, to test code and pipelines.