IVK-395 Add testcase for seqid-bitscore, seqid-evalue extraction and sorting
The .CSV file output by blast only contains bitscores and evalues for each HSP alignment, of which there can be many for each seqid.

The .HTML blast outputs, includes the calculated rollup of those (possibly overlapping) alignments to a full-sequence (seqid) level bitscore and evalue.

The results in both file are sorted by ```seqid_bitsore DESC, hsp_bitscore DESC```

We want to find a way to capture the seqid_bitscore and put in in the JSON. We *can* extract it from the HTML, but it would be better if we could get it from XML or another format that's safer to parse.